App component moved to src/Components/App folder -- all imports edited -- site working
