+++ src/Components/App/index.js
@@ -1,5 +1,5 @@
-import logo from './logo.svg';
-import './App.css';
+import logo from "../../logo.svg";
+import "./App.css";
 
 function App() {
   return (
@@ -13,8 +13,7 @@ function App() {
           className="App-link"
           href="https://reactjs.org"
           target="_blank"
-          rel="noopener noreferrer"
-        >
+          rel="noopener noreferrer">
           Learn React
         </a>
       </header>
